[backend] Add test for KanbanCsv
